# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bac Thinh API docs",
    "description": "Mo ta API service",
    "license": {
      "name": "API License",
      "url": "http://nguyenkiem.vn/license"
    },
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Server test"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/weekly-schedules/{id}": {
      "get": {
        "tags": ["Weekly Schedule Management"],
        "summary": "Get weekly schedule by ID",
        "description": "Retrieve a specific weekly mass schedule",
        "operationId": "getWeeklyScheduleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Weekly schedule found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Weekly schedule not found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Weekly Schedule Management"],
        "summary": "Update a weekly schedule",
        "description": "Update an existing weekly mass schedule",
        "operationId": "updateWeeklySchedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWeeklyScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Weekly schedule updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Weekly schedule not found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Weekly Schedule Management"],
        "summary": "Delete a weekly schedule",
        "description": "Delete a weekly mass schedule",
        "operationId": "deleteWeeklySchedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Weekly schedule deleted successfully"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Weekly schedule not found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/weekly-schedules": {
      "post": {
        "tags": ["Weekly Schedule Management"],
        "summary": "Create a new weekly schedule",
        "description": "Create a new weekly mass schedule",
        "operationId": "createWeeklySchedule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeeklyScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Weekly schedule created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Weekly schedule already exists for this date range",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weekly-schedules/{id}/publish": {
      "post": {
        "tags": ["Weekly Schedule Management"],
        "summary": "Publish weekly schedule",
        "description": "Publish a weekly schedule to make it official",
        "operationId": "publishWeeklySchedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Weekly schedule published successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Weekly schedule not found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weekly-schedules/{id}/archive": {
      "post": {
        "tags": ["Weekly Schedule Management"],
        "summary": "Archive weekly schedule",
        "description": "Archive a weekly schedule",
        "operationId": "archiveWeeklySchedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Weekly schedule archived successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Weekly schedule not found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/weekly-schedules/search": {
      "get": {
        "tags": ["Weekly Schedule Management"],
        "summary": "Search weekly schedules",
        "description": "Search and filter weekly mass schedules",
        "operationId": "searchWeeklySchedules",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "weekNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "PUBLISHED", "ARCHIVED"]
            }
          },
          {
            "name": "createdBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "weekStartDate"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DESC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search completed successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedWeeklyScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weekly-schedules/by-date": {
      "get": {
        "tags": ["Weekly Schedule Management"],
        "summary": "Get weekly schedule by date",
        "description": "Get weekly schedule containing a specific date",
        "operationId": "getWeeklyScheduleByDate",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Weekly schedule found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "No weekly schedule found for this date",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weekly-schedules/by-date-range": {
      "get": {
        "tags": ["Weekly Schedule Management"],
        "summary": "Get weekly schedules by date range",
        "description": "Get all weekly schedules within a date range",
        "operationId": "getWeeklySchedulesByDateRange",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Weekly schedules found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeeklyScheduleResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UpdateDailyScheduleRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dayOfWeek": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "masses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateMassRequest"
            }
          }
        },
        "required": ["date", "dayOfWeek"]
      },
...
      "UpdateWeeklyScheduleRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "dailySchedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateDailyScheduleRequest"
            }
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "PUBLISHED", "ARCHIVED"]
          }
        }
      },
      "DailyScheduleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dayOfWeek": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "masses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MassResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "WeeklyScheduleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "weekStartDate": {
            "type": "string",
            "format": "date"
          },
          "weekEndDate": {
            "type": "string",
            "format": "date"
          },
          "weekNumber": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "dailySchedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyScheduleResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "PUBLISHED", "ARCHIVED"]
          }
        }
      },
...
      "CreateWeeklyScheduleRequest": {
        "type": "object",
        "properties": {
          "weekStartDate": {
            "type": "string",
            "format": "date"
          },
          "weekEndDate": {
            "type": "string",
            "format": "date"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "dailySchedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateDailyScheduleRequest"
            }
          },
          "createdBy": {
            "type": "string"
          }
        },
        "required": ["weekEndDate", "weekStartDate"]
      },
...
      "PagedWeeklyScheduleResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeeklyScheduleResponse"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}